Changes to ElasticSearch to use Three tiered style search
Search has been designed to be a three tiered approach to optimize
certain business and search criteria, and has been broken down into the
following rules:

1. Exact Match to RegIdentifier
2. Exact match to any other value
3. Fuzzy match to certain fields, such as company name, postcode, first
and last names

Each of these queries will run in turn and if any tier finds any number
of matches, those matches are returned.

For example if a RegIdentifier is specified and an exact match is
found, only that one result is returned.
If However a match is not found the search continues onto the next
tier, i.e. Exact match to any other value, and so on.