Marked eclipse .launch files as ignored.
Git keeps wanting to check in my WasteCarrierService.launch file, but
as this is eclipse related and machine specific it’s of no use to
anyone else. Having told GitHub for Mac to ignore it, it updates the
.gitignore file and sets the .launch file as deleted, which need to be
checked in and synced to allow me to pull down the latest versions of
the actual code.

Really hope this doesn’t ruin anyones day!