Amended public register search sort order
Users were finding that when searching for something specific such as
‘crofty nurseries gardening services’ even if a record existed with
that exact title, it would not appear at the top of the results or even
possibly at all.

After some testing found it was because we were sorting by name, and
ignoring the actual score Elasticsearch had assigned to each result. We
could match users expectations if we ensured that the result with the
highest score was shown 1st, and any with the same score were sorted in
alphabetical order.