Fix issues and formatting in README
* Update URL, fix Markdown formatting

* Remove ver. no. in elasticsearch section

The version stated was 0.90.5 which is out of date. Rather than keep needing to update the version to reflect what we have deployed, this change simply removes the version from the details in the README.

* Correct license details

The README was previously stating that the license is MIT. The EA has agreed the use of the Open Government License or OGL for all its projects hence this correction.

Also added the **Contributing to this project** to bring it more inline with other projects.