Add find regisration by reg. number feature
https://eaflood.atlassian.net/browse/WC-84

To support changes to the front end to properly manage users entering new style registration numbers when trying to renew, we need to be able to confirm if the number entered actually exists.

Surprisingly there was no ability to search for a registration using just its registration number previously. Hence this change adds in support for that kind of search.