Add find registration by orig reg. no feature
https://eaflood.atlassian.net/browse/WC-84

To support changes in the front end to make the renewals process more robust, we need to be able to search registrations by `originalRegistrationNumber`. Specifically this will help us block a look hole where the same IR registration number can be used multiple times to create new registrations with the discounted £105 charge.