Add find registration by original registration number feature (#3)
https://eaflood.atlassian.net/browse/WC-84

To support changes in the front end to make the renewals process more robust, we need to be able to search registrations by `originalRegistrationNumber`. Specifically this will help us block a look hole where the same IR registration number can be used multiple times to create new registrations with the discounted £105 charge.

* Ensure .log files are ignored

Whilst coding the search by original registration number the file `hs_err_pid2179.log` was generated in the root of the project. To ensure these don't get added to source control this change updates the .gitignore file to ignore them.

* Fix most glaring issues in the README

The README probably still needs some work to properly document all the features of the project, however this change resolves the most glaring issues in it which were starting to bug me!