Update test IR data with valid active Comp. reg No (#4)
The Company registration number being used in the test IR data import file now belongs to a company that has been liquidated. The frontend feature tests however expect the number to belong to an active limited company and hence are failing.

This change replaces the number in the file with one known to be active.