Update fixture to fix tests on wcr-vagrant-36 box
As part of the migration to MongoDb 3.6 we have built a new vagrant box that runs Ubuntu 16.04 and MongoDb 3.6. The first issue we faced when compiling the project was that the serialisation tests failed due to new changes in the actual outcome. This change updates the expected result (set out in the fixture) to match the actual.