Upgrade mongodb driver and mongojack POJO mapping
This change updates the version of the MongoDb driver to the latest that supports version 3.6, as well as the latest version of Mongojack.

It also deletes what is in fact a duplicate reference to Mongojack. Back in 2013 they [changed the package name](https://github.com/mongojack/mongojack/commit/15bf67885f3b895fdc8a4cecce78817acc2895d4) so a reference to `net.vz.mongodb.jackson` was just a reference to an older version of the library.