Refactor DBHelper to use new MongoDb driver v3 API
In version 2 of the MongoDb driver a connection would be made with the MongoClient, passing in the host and port. Then a separate call would be made to authenticate the connection before then returning a DB object for use elsewhere in our code. Wherever the DB object was used the calling code would then have to check its was authenticated.

In version 3 they have removed the separate authentication and checking mechanism and now you authenticate as part of instantiating the connection.

It should be noted that the use of DB, and things like DBCollection have been deprecated in favour of MongoDatabase and MongoCollections. We would have refactored to also use them (in fact we did start) however we also rely heavily Mongojack and its `JacksonDBCollection` to do things. If you look at their documentation for the latest version they are still covering this as the main way to query MongoDb with their POJO mapper, though they have now included support for the [new version 3 API](https://github.com/mongojack/mongojack#using-jacksonmongocollection). However this would require an extensive rewrite of the code so for now we accept we are using deprecated objects and method calls when it comes to the DatabaseHelper class.