Refactor MongoHealthCheck to use DatabaseHelper
Prior to this chance we were passing in a MongoClient instance and then letting the class do all the same stuff as the database helper, plus requiring a hard coded database name. Now with the refactoring the class uses DatabaseHelper like most others, and the reliance on a hard coded database name is gone.