Refactor Reg.StatusJob to support MongoDb 3.6
Change similar to previous refactorings, though in this case we still handle an error when updating the registrations.

Based on the knowledge that `WriteResult.getError()` should not be used and instead you can just fall back on the exception thrown for details we have been removing any checks for `getError()` not returning null as in almost all cases we throw an exception anyway.

In this case we seem to swallow it, hence we have wrapped the attempt to update the db in a try catch so we catch any exceptions and handle them in the same way as before.