Refactor LocationPopulator to support MongoDb 3.6
Changes similar to previous refactorings, though we have explicitly handled in a try/catch a write that errors (unlike in some other cases).

We did it in this case because the previous logic appears to throw an exception with a specific status. In case there is logic depending on this we have used the try/catch to do the same.