Update fixtures to match actual
Having made all the changes ti support MongoDb 3.6 then found that though we could compile we could not run because of java.lang.UnsupportedClassVersionError error.

It looks like having updated our packages we now need to be compiling and running with Jave 8. This was confirmed when having removed Java 7 from the development machine we were able to move past the java.lang.UnsupportedClassVersionError error. However it looks like what is now output following the serialization of the object matches what was expected at the very start, hence we need to update the fixture back to what it was.