Force use of later jackson dependency
Having confirmed we were connecting to MongoDb successfully at this point, we found that we could not get MongoJack and Jackson to bind to the Registration class. Specifically

```java
JacksonDBCollection<Registration, String> registrations = JacksonDBCollection.wrap(
					db.getCollection(Registration.COLLECTION_NAME),
					Registration.class,
					String.class
			);
```

Some searching later and the reason appeared to be an incompatibility between the version of fasterxml.jackson.core being used by the project being very old. Adding it as a specific dependency appeared to solve the problem for MongoJack, and the code above plus queries to the database started working again.