Reinstate MongoDb connection test
In order to first diagnose and confirm that we were connecting successfully to MongoDb 3.6 following the upgrade we made use of the existing MongoDb test.

Whilst we are still in the throes of upgrading the project to support MongoDb 3.6 this change adds the test back in, so we can at least confirm its not a connection issue when future issues arise.

Unfortunately because we haven't implemented a way to read a local config file, its does depend on environment variables to exist when `mvn package` is run. To aid this we have added a bash script which will source them for you, and add as a guide to what's needed. The details in the file match what is used in our Vagrant environment.