Upgrade Dropwizard to 1.3.1
Having confirmed that following the upgrade to the MongoDb driver to the latest version, and MongoJack likewise that we were able to connect and query MongoDb, we then were getting errors when calling the API.

More searching found this was because of incompatibilities between the dependencies these libraries were now using, and those in the existing Dropwizard libaries. The only recourse appeared to be to update Dropwizard to the latest version as well.

Seeing as we were going from version 0.6.2 (at time of writing the oldest version they list!) to 1.3.1 large scale changes were expected in the API. Hence necessary changes throughout the files.

All changes here are in support of getting project updated and compiling.