Delete pointless and annoying test
Whenever we make a change to the project it seems to then require a change to the test fixtures because they no longer match how registrations are actually serialised. Plus you could update the fixtures and get the tests passing, but the app itself could be completely broken!

The update to Dropwizard 1.3.1 meant a number of helpers that the test relied on were no longer available. Rather than work through trying to find alternatives we finally took the decision to just dumpo the test.

And good riddance!