Update logback dependency
When attempting to run the app was getting an error related to ch.qos.logback.classic (Failed to instantiate [ch.qos.logback.classic.LoggerContext]). Again searching for answers talked about incompatibilities between the version of the logback library used by dependencies in the project.

So using `mvn dependency:tree` best guess was that airbrake-logback was using bringing in a pretty old version. However even updating it still didn't resolve the issue hence next recourse is to define it directly in the pom.xml file.