Expand current MongoDb test coverage
We wanted to provide support for quickly testing that the service can connect and manage data in MongoDb. So we have gone with keeping in the existing tests to check the connection and that we can create a registration, and expanded them to include find and update.

We have also updated the test class to use the annotation format (which is the current model of doing JUnit tests) and to support all the tests whilst avoiding duplication configured them to run in a specific order with a clean down for the registration created.

However to avoid needing to have an environment where MongoDb, the databases and the collections exist, we have configured the pom so the tests are no longer run during a package. They now need to be invoked manually, but this fits better with there intended purpose of checking the services can connect rather than that the code is behaving as expected.