Remove redundant No. pending registrations code
We had a test for a method that tells us how many pending registrations there are. As soon as we removed it, it highlighted that the code itself was only used by the test. So this change removes both the test and code from the solution as it is redundant.