Add support for Maven wrapper
https://github.com/takari/maven-wrapper

Maven wrapper ensures that any user of your project can build it, without first worrying about installing Maven. The project comes with scripts that will download and install Maven if it detects that Maven is missing when you run it.

The primary reason for adding it is to support deployment of the service into a new AWS environment, where deployment is managed by Jenkins. By making use of the Maven wrapper we don't need to ensure that the Jenkins instance has Maven already installed when we come to build the project.