Add Jenkins pipeline script to the project
https://eaflood.atlassian.net/browse/WC-157

This script is used by our Jenkins instance to build the project. Using pipeline support built into Jenkins it means we can define how the project should be built, and the steps to take in that process in a way that is codified and source controlled.

Essentially a pipepline job is then created in Jenkins that points to this repo, and which will recognise and pick up this script. That is essentially it. No further configuration is needed in the job as it takes its instructions from the script.