Add ignore unknown attribute to all objects
It seems with the upgrade to the latest version of Dropwizard that if we pass data in the Json from the front end to the back end that does not exist in ther matching POJO that it causes an error (specifically a http 400 bad request error).

We haven't looked into the release docs, but our guess is that somewhere along the way either the functionality around the `@JsonIgnoreProperties(ignoreUnknown = true)` was strengthened, or old behaviour of inheriting this from a parent object was stopped.

Either way. This seems to resolve an issue that has only arisen since upgrading the services.