Swallow error when no user found in getUserByEmail
Whilst attemping to get all the feature tests passing in the waste-carriers-frontend project found we were hitting an error whenever attempted to update a registration.

After some debugging tracked down the cause as the services were erroring in `UsersMongoDao.getUserByEmail()`. At that time in the test the user had not been created, so this led us to double check whether throwing an error was the expected behaviour.

After referring back to what the method looked like before our MongoDb 3.6 changes we did indeed find that in the event of no user being found we simply return null rather than throw an error.