Incorporate env vars into the configuration.yml
https://eaflood.atlassian.net/browse/WC-157

To improve how the service can be deployed rather than have to hard code various values into a config file (specifically database username and passwords) and then store that file elsewhere, we've taken advantage of the upgrade in the Dropwizard framework to use environment variable substitution.

Depending on where the app is deployed to (either vagrant or an AWS environment) you'll need to set certain environment variables to match as some of the defaults are for a specific environment.

What it does now mean though, as long as those environment variables are set by the environments provisioning scripts, we can start the service with the `configuration.yml` file included in the project.