Update repurpose reindex_elasticsearch.sh
The old version of this file would do somthing with the postcodes stored in postcodes.csv (this is still to be investigated) and then run the reindex task, which essentially dumps what elasticsearch currently has and then re-adds each registration one by one.

As we removed that task in a previous commit, the only reason for this file to exist is to handle the postcodes hence the update and change of name.