Remove redundant deployment related artifacts
https://eaflood.atlassian.net/browse/WC-157

When it comes to deploying this service into our new AWS environment we'll be using Jenkins pipeline feature. Hence we have a `Jenkinsfile` which handles building the app, and another saved in the Waste carriers Jenkins repo that will handle deploying the built artifacts, setting any env vars and then running it.

Because of these changes the bash scripts contained in the project are redundant so this change deletes them to remove any confusion in the future.