Remove dependency on Elasticsearch
https://eaflood.atlassian.net/browse/WC-309

As part of simplifing the waste carriers solution ready for its migration to AWS we are removing Elasticsearch.

There are no notes from the time, and none of the current team were present when it was decided to use it. So our assumptions are that the original team thought this would provide more intelligent search functionality when it came to users looking for a waste carrier.

We certainly don't believe there were any specific user needs it was catering for, and now with the public register gone the only place its used in the service is when an internal user first searches for a registration.

However to support this we essentially have to duplicate every registration in our main database (MongoDb 3.6) into Elasticsearch. This means any time we update a record in MongoDb, we also need to make the same change to the record in Elasticsearch.

All this means lots of additional code and complexity, when we could simply just search for registrations using MongoDb directly. It also simplifies any fixes or changes we want to apply in the future, as we no longer have to also account for managing what happens to the data in Elasticsearch.

So because of all these reasons, when we launch the service anew in June in AWS we do not intend to have Elasticsearch as part of it. The first step in achieving this is to remove it from the Waste carriers service so that it no longer depends on Elasticsearch being installed and running in order to be able to start up and function.

WIP