Remove Databasecleaner task
This task does clear the MongoDb database of registrations but also has a heavy dependency on Elasticsearch as it does the same job there.

However when working locally we now have scripts that can much more efficiently clear the database, and when running our tests we use the gem [DatabaseCleaner](https://github.com/DatabaseCleaner/database_cleaner).

So rather than rework this task to remove Elasticsearch, we're just deleting it as its no longer used.