Remove Elasticsearch from PaymentResource
Payment records can be fired at the waste carriers service separately from a registration and as such need to be attached to the relevant registration.

As we have to duplicate everything in Elasticsearch, it would have also updated the registration there. This change removes all Elasticsearch related code for doing that.