Refactor search for reg. by accountEmail
This change refactors the existing function of searching by accountEmail for registrations. It moves the end point to the QueryResource, and changes the code in the search to make use of MongoJack rather than build the query using just the underlying MongoDb driver.

The original was written when we had less understanding of where MongoJack/JacksonDb fitted within the solution. We now understand we probably should be using it for queries unless we need to drop down to the underlying MongoDb driver for a specific reason.

To do this meant opening/adding methods to the SearchHelper, something itself that could probably do with a refactor.