Refactor unit tests and expand reg. builder
All the tests require access to the environment variables in order to make a connection to MongoDb. Plus when the tests are done all need to clear out any they created registrations.

Because of this we have moved this code to a new shared class called `ConnectionsUtil` which from now on will handle these details.

On top of that in preparation for adding more tests we have expanded the functionality of the `RegistrationBuilder` so that now we can create either a lower or upper tier registration, or even an upper tier IR renewal regiustration.