Update test env vars to use test database
So that we are working with known registrations in the unit tests, each test class will create the registrations it needs then empty (actually delete) the collection once complete.

Because of this pointing to the actual database means it'll be wiped, which can be a pain if it contained registrations you wanted to keep.

So in future there is an assumption that a test database is available and so this change updates the env vars we set before running the tests.