Refactor CopyCard search
This refactors the copy card search plus adds some unit tests for it.

Some of the refactoring comes from moving things into the new search namespace, but the majority comes from switching to using MongoJack for the querying, rather than going direct to the underlying MongDb driver.

This switch simplifies the queries and makes them much more readable, in order to understand what they are doing.

We've have also checked what is going on from the frontend, and rather than coding for all possible scenarios, we have stripped the functionality/behaviour to match only what the front end requires.

Finally we deal with some of the repurcussions of the name change of `QueryResource` to `SearchResource`.