Refactor tests plus only create test data once
Refactored the tests to make them more consistent, plus the tests will work with all the test data being created upfront so have moved the creation into the the @BeforeClass.

Also settled on a naming convention of `setup()` and `tearDown()` for methods marked `@BeforeClass` and `@AfterClass`.

Finally fixed a mistake in RegistrationBuilder where payments were not being added to copy card registrations.