Fix wrong use of IRRENEW order type
Whilst working on a new unit test spotted that the `RegustrationBuilder` is using the wrong order item type when it comes to IR Renewals. Based on code in the front end and the comment in `OrderItem` the type `IRRENEW` is not actually used (and I can't find any use in the code).

So this change both removes it and fixes `RegistrationBuilder`.