Make copycard distinct from BuilDType in RegBuildr
Started out with the intention of having build types lower, upper, upper copy and renewal. But looking ahead to the need to add a renewal_copy and the extra this would add to all the switch statements we already have, decided instead on a refactor.

So this change returns build type to just lower, upper, and renewal, and adds the need for a copy card aswell as an extra parameter in the constructor. This is then fed through the class by updating the switch statements based on build type, and adding logic for copy cards where appropriate.

N.B. Apologies the commit is a little messy. The changes were made during a larger refactor however wanted to capture them as a distinct change so as not to lose the meaning.