Reword should not include assertions in tests
Whilst working on a new test rather than leave the expectation as being `true` and then using an alternative to `anyMatch()` (`noneMatch()`) to get the result I switched it.

So the expectation was changed to `false`, and we switched to `anyMatch()` to get the actual result.

Though a minor change, I believe it reads better plus it means we now use the same syntax to determine whether a value exists in our results, irrespective of whether the expectation is it should or should not.

Much easier copying & pasting!