Refactor main search in RegistrationsResource
Currently the `getRegistrations()` in the `RegistrationsResource` class is a massive 500 line behemoth that is tightly coupled to Elasticsearch.

However on review a lot of the logic is no longer used, mainly due to the removal of the public register from the service. The key thing it still supports is the search internal users use to find a users registration, accessed from the home screen of the back office app.

To make the search consistent with others we have been refactoring, this change creates a new `WithinSearch` class that with changes to the frontend will now be the basis for that internal search.

We can then begin the work of removing the old code once we have tested the search from the front end.