Update test db env vars to remove conflicts
This change updates the name of the env vars the unit tests rely on when they run, along with the helper file which can set them when working locally.

Realised that if we set the env vars, we may overwrite others we have set for working in a development mode. Hence changed the tests to instead use new env vars to get the MongoDb connection details, and updated the `test_env_vars.sh` to match.