Add default search result count as config setting
Rather than just hardcoding in a set number of values to return, we added logic to instead read from an environment variable, though we do provide a default of 50.

In this way should 50 be deemed not a suitable number, then we can change it with just an env var change and a restart of the services.