Remove remaining function that made use of elastic
This change removes the massive 400+ line function that used to back the main back office registration search page. We've now moved what we believe the front end needs to work to `WithinSearch` and have deleted the rest.

With this change all remaining references to Elasticsearch should have been removed from the project.