Fix for broken IR reg number search end point
When trying to get the frontend talking to the services via `SearchResource.queryOriginalRegNumber()` quickly realised that we kept getting http 400's.

After some trial and error we realised that the issue was the use of camel-case for the endpoint name. We had defined the path as `/originalRegistrationNumber/` before but even if we made sure to use the right case, it was not connecting.

Hence this corrects issues with the definition of the endpoint to ensure it works.