Add new `DatabaseHelper.getCollection` + refactor
Our first entity matching test will be one that simply confirms we can connect to the database. To do that we need to access the main collection in it.

The current `SearchHelper` contains a `getRegistrationsCollection()` from which we can crib how to get the collection, however `SearchHelper` is specific to registrations.

So to have a more reusable option we have created a `getCollection()` method on the `DatabaseHelper` which takes the name of the collection and returns it. As a `DatabaseHelper` is instantiated with a `DatabaseConfiguration` it already knows how to connect to the selected database.

With this new method we then refactor some of the existing code to make use of it, and to remove some unnecessary methods.