Refactor SearchHelper and RegistrationDao
The only thing that was making the SearchHelper unique to Registrations was its `registrationsCollections` method.

It could be useful when we come to do the Entity searching, so the first step in refactoring was to make that method generic (plus rename it).

Then rather than reinvent the wheel we wanted to make use of the `getCollection` methods on the dao objects (at this point it existed on the EntityMatchingDao but not the Registrations). To do this we need to be able to pass in Dao 'like' objects that returned a generic JacksonDbCollection. Hence we added a new interface and had both Dao objects implement it.

With that done we then made some further refactorings to make the Entity and Registration dao classes more consistent, hence renaming the methods (and the subsequent changes that caused).

Finally we updated the (now) RegistrationDaoTest so that the individual tests not longer relied on each other.