OriginalRegNumberSearch should only return 1 reg.
Obviously a result of being to aggresive with the cut and paste, prior to this change `OriginalRegNumberSearch` was returning a list of registrations.

However the calling method in the front end is only ever expecting one result hence was erroring.

This change updates the search to match the expected behaviour.