Migrate entity check functions to services
https://eaflood.atlassian.net/browse/WC-342
https://eaflood.atlassian.net/browse/WC-83
https://eaflood.atlassian.net/browse/WC-158

WC-83 is a story about updating the matching algorithm in what is currently called the convictions service. This essentially means a re-write of the current algorithm.

WC-158 is about the need to automate the deployment of the convictions service ready for the move to our AWS environment.

Finally WC-342 is about removing the dependency on Elasticsearch within the service in order to simplify it, making it easier to maintain in the future.

The solution we have come up with to these is to migrate the existing the functionality in the convictions service into this project.

Elasticsearch has already been removed, plus it has been update to the latest Dropwizard framework and made to work with MongoDb 3.6 already (where we'll need to store our data for this service in future).

Doing this reduces the complexity of the service, we have one less app to worry about. We won't need a to create a separate automated deployment process as once we've created the one for this app, we're done. Plus as both services use Java we can more easily migrate the code.

Hence this PR covers all changes required to

- Migrate the existing functionality
- Update it to work with MongoDb 3.6 instead of Elasticsearch as its datastore
- Update the algorithm to match what has been agreed with the business and the NCCC team