Update PersonMatch to handle 'LN, FN' and 'FN LN'
The previous version of the version match worked on the assumption that all names in the entities collection would be stored as `Lastname, Firstname`. However a review of the code confirmed this is not always the case and they can also be stored as `Firstname Lastname`.

As the updated algorithm expects us to onbly return results where both the first and lastname are found we needed to tweak the code to handle this.