Implement a CSV reader for the entities file
Entity details to load into the system are provided by CSV file. This change adds a new `EntityCsvReader` class which understands the format of the file we can expect.

Other changes are because also took the opportunity to update openCSV which is used to simplify the reading of CSV files.