Add new method to EntityDao to recreate collection
With entities the expectation is periodically we will be provided a fresh list of the current entities to match against.

Hence when populating the database, we drop and recreate the collection. This change implements that ability.

It also renames the `EntityMatchingDao` to bring it inline with the other dao's. Essentially they are matching to single collections not databases, hence the change in the name of this one.