Add server side selection config to db connections
In preparation for adding some new functionality around checking connections and the unit tests for them this change adds the ability to configure and set the server side selection timeout property on the MongoClient created for a connection.

In a unit test under development we wanted to check that an exception is thrown when the database credentials are invalid. The test works, but because the default timeout for this property is 30 seconds, we have to wait that amount of time before the test fails.

Having this configurable property allows us to set it to a much smaller time when running unit tests, but we leave the default (and don't expect to use it) for when running in production.