Remove MongoManaged
The idea of MongoManaged seems to come from the fact you would create a MongoClient and call close on it when the service closes. A recent blog post I found that demonstrates putting Dropwizard and MongoDb reiterated this strategy.

The problem is at some point we moved the creation of the MongoClient into the Databasehelper as we need to deal with 3 different databases.

Therefore we have overlooked these over connections, without it appears any detrimental effect. Therefore this change removes it entirely.