Update EntityPopulator task to handle json
This change makes the EntityPopulator task multi purpose. Now if called with no message body it will re-populate the entities collection from the configured csv file.

But if the POST body contains a JSON array of entities it will instead use that. This feature was specifically added to support acceptance and performance testing, which previously have found it difficult to specify test data appropriate for their needs.