Fix issue with the BackgroundJobScheduler
Since adding the new `serverSelectionTimeout` config we have been seeing errors in the console in relation to the background jobs specifically `Identified object is not an integer.

We then spotted that we were not passing the info from the database config into the background jobs config hence the error. This change fixes it.