Remove getters and setters from ConSearchResult
As part of starting to bring together the entity matching and convicton search objects, this change takes the opportunity to remove the getters and setters from the class.

Possibly a controversial move, but we have decided that where a class property has a getter and setter that perform no computation, they become instead noise and unnecessary code and hence should be removed.