Fix broken test and code in CompanyMatch
Found we actually had problems with the company match.

Because of the test data we are now seeding it was returning a different result in one test hence we needed to update it to match.

The `numberMatchFuzzy` was failing for valid reasons however. We thought we had the regex setup in the `numberMatchIgnoresZeroes` method properly i.e. match the search value against the end of numbers in the db but we were wrong.

This change fixes the code and ensures the test is matching our expected behaviour (previous search value should have been a match as it us the end of a number).