Fix broken PersonMatcher & include DOB in resource
When coming to add date of birth as an arg to the person matching end point in `MatchResource` realised that we needed to pass in a string to `PersonMatch` and have it convert it to an actual date object.

With then done then spotted that our tests were not actually working. They were returning a match but not one where the dates actually had been used.

This therefore meant the need for a fix to the PersonMatch query to ensure that when a Date of birth is specified and a matching record exists it is picked in favour of one where just the name matches.