Fix error found in search resource
Found an error in the `SearchResource` with endpoints that were using path params for values that were being encoded in the front end.

So for example `queryWithinI()` prior to the change had a path of `/registrations/{searchWithin}/{searchValue}`.

The problem was that if you searched for `Hello Kitten` it was being escaped as `Hello+Kitten` and `queryWithin()` was not removing the '+' from the param. Switch it to a query param however and it handles it properly.

So this change updates any path param to a query param where the value is likely to be escaped.