Add exported_data to .gitignore
If you want to run the service locally but not get errors in the console from the background job, you must provide a valid config value for `eprExportFile` and `reportingExportPath`.

Generally this means just creating a folder at the root of the project called 'exported_data'. Hence adding it to the .gitignore.