Handle names with periods in CompanyMatch
https://eaflood.atlassian.net/browse/WC-342

During testing of the new entity matching features it was highlighted that querying for 'Test Waste Services Ltd.' didn't match with a 'Test Waste Services Ltd'.

It was all down to the end period, and to guard against it in future the `CompanyMatch` now strips any periods of the end of names passed in.