Remove unused postcode to lat-long functionality
https://eaflood.atlassian.net/browse/WC-309

A missing left over from the services reliance on Elasticsearch. The services included a feature where each time a registration was added a big csv file of postcode to lat/long conversions was searched, and the results added to the registration.

This information could then be used in the early versions of the public register when attempts were made to search within a given radius.

As we don't really use this info anymore, and it comes with an address when we use the address lookup, this change removes it to reduce the maintenance and complexity of this project.