Fix error with background job config
In further testing following the env var standardisation we spotted that the `DatabaseConfiguration.getUrl()` method was actually returning the url property.

This was then causing an error when the background jobs were being scheduled.

This change fixes it. We also add back in references to host and port to the jobs data maps just in case.