Update worldpay mock to support frontend (#27)
https://eaflood.atlassian.net/browse/WC-365

When testing with the frontend some additional issues arose. It doesn't send through a URL for all possible situations so we have had to amend the validations on the resource action.

Also the url's it does send through already include a query param so we had to amend how we build the redirect url.