Add ability to mock companies house to services
https://eaflood.atlassian.net/browse/WC-408

These changes add the ability for the waste-carriers-service to 'mock' Companies House.

In cases where we need to carry out performance testing we don't want to hit the actual services, for example OS Places for address lookup and Worldpay for payments.

We could have built a separate companies-house-mock-app but then that would mean deploying an extra app just for when we need to do performance testing.

By building it into the service, we simply need to change the env var which holds the url for Companies House to point to our local instance, and we are able to begin mocking it.