Remove need for entity-matching database
When we migrated the entity matching functionality from a stand alone app into this existing services project, we kept the design of it using its own database.

As part of the drive to further simplify the waste carriers service, we came to the conclusion that it could be maintained more easily as just another collection in the existing waste carriers database (referred to as the 'registrations' database in the code).

Hence this change strips out configuration and logic which previously setup or depended on a separate entity-matching database.